requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
PyYAML>=6.0.0
dash>=2.17.0
plotly>=5.20.0
tqdm>=4.66.0
